create author facade
