+++ src/facades/book.js
@@ -2,16 +2,6 @@
 
 let books = require('../../db.json')
 
-function getAuthors () {
-  let result = new Set(
-    books
-      .map(book => book.authors)
-      .reduce((prev, curr) => prev.concat(curr))
-      .sort()
-    )
-  return [...result]
-}
-
 function getBooks (query) {
   let result = books
   if (query.category) {
@@ -39,6 +29,5 @@ function getBooksById (id) {
   return books.find(b => b.id === id)
 }
 
-exports.getAuthors = getAuthors
 exports.getBooks = getBooks
 exports.getBooksById = getBooksById
